주식 완성 + 상점 페이지 만듬

+++ Assets/Scripts/ItemCard.cs
@@ -14,7 +14,7 @@ public abstract class ItemCard : MonoBehaviour
     protected TextMeshProUGUI desc;
     protected TextMeshProUGUI buttonText;
 
-    public int BuyMoney;
+    public long BuyMoney;
     protected virtual void Start()
     {
         Image image = gameObject.transform.Find("Icon").GetComponent<Image>();
